Add config and load token from file

+++ src/main/java/com/smyhktech/bot/tutorial/objects/CommandManager.java
@@ -64,7 +64,9 @@ public class CommandManager {
                 if (command.getLabel().equalsIgnoreCase(label)) {
                     worked = true;
                     command.runCommand(user, channel, guild, label, args);
-                    event.getMessage().delete();
+                    if (!channel.isPrivate()) {
+                        event.getMessage().delete();
+                    }
                 }
             } // end for
